Int - VarInt separation

+++ src/main/kotlin/olivermakesco/de/kraft/network/NetworkState.kt
@@ -1,6 +1,6 @@
 package olivermakesco.de.kraft.network
 
-enum class NetworkState(val id: Byte) {
+enum class NetworkState(val id: Int) {
     HANDSHAKING(-1),
     PLAY(0),
     STATUS(1),
